prepublish commands, move tests to separate dir
