+++ .npmignore
@@ -1,3 +1,5 @@
 node_modules
-.circleci
 *.tgz
+coverage
+src
+test
