+++ vitest.config.ts
@@ -2,10 +2,6 @@ import { defineConfig } from 'vite';
 
 export default defineConfig({
   test: {
-    includeSource: ['src/**/*.{js,ts}'],
-    coverage: {
-      include: ['src/**/*.ts'],
-      exclude: ['src/index.ts']
-    }
+    coverage: {}
   },
 });
